requirements.txt
pandas
numpy
pyyaml
pytest
python-dotenv
pydantic
loguru
scikit-learn
jupyter
openai  # if you use LLMs